samquotes go brrrrr
