--- src/main/java/com/diamondfire/helpbot/bot/command/argument/impl/types/EndlessStringArgument.java
@@ -1,15 +1,0 @@
-package com.diamondfire.helpbot.bot.command.argument.impl.types;
-
-import com.diamondfire.helpbot.bot.command.argument.impl.parsing.exceptions.ArgumentException;
-import com.diamondfire.helpbot.bot.events.CommandEvent;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.Deque;
-
-public class EndlessStringArgument implements Argument<String> {
-    
-    @Override
-    public String parseValue(@NotNull Deque<String> args, CommandEvent event) throws ArgumentException {
-        return String.join(" ", args);
-    }
-}
