+++ src/main/java/com/diamondfire/helpbot/bot/command/argument/impl/types/GreedyStringArgument.java
@@ -1,4 +1,4 @@
-package com.diamondfire.helpbot.bot.command.argument.impl.parsing.types;
+package com.diamondfire.helpbot.bot.command.argument.impl.types;
 
 import com.diamondfire.helpbot.bot.command.argument.impl.parsing.exceptions.ArgumentException;
 import com.diamondfire.helpbot.bot.command.argument.impl.types.Argument;
@@ -7,7 +7,7 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.Deque;
 
-public class MessageArgument implements Argument<String> {
+public class GreedyStringArgument implements Argument<String> {
     
     @Override
     public String parseValue(@NotNull Deque<String> args, CommandEvent event) throws ArgumentException {
