+++ src/main/java/com/diamondfire/helpbot/bot/command/impl/other/dev/EvalCommand.java
@@ -2,7 +2,7 @@ package com.diamondfire.helpbot.bot.command.impl.other.dev;
 
 import com.diamondfire.helpbot.bot.HelpBotInstance;
 import com.diamondfire.helpbot.bot.command.argument.ArgumentSet;
-import com.diamondfire.helpbot.bot.command.argument.impl.parsing.types.MessageArgument;
+import com.diamondfire.helpbot.bot.command.argument.impl.types.GreedyStringArgument;
 import com.diamondfire.helpbot.bot.command.help.*;
 import com.diamondfire.helpbot.bot.command.impl.Command;
 import com.diamondfire.helpbot.bot.command.permissions.Permission;
@@ -46,7 +46,7 @@ public class EvalCommand extends Command {
     public ArgumentSet compileArguments() {
         return new ArgumentSet()
                 .addArgument("code",
-                        new MessageArgument());
+                        new GreedyStringArgument());
     }
     
     @Override
