+++ src/main/java/com/diamondfire/helpbot/bot/command/impl/other/fun/EightBallCommand.java
@@ -2,6 +2,7 @@ package com.diamondfire.helpbot.bot.command.impl.other.fun;
 
 import com.diamondfire.helpbot.bot.command.argument.ArgumentSet;
 import com.diamondfire.helpbot.bot.command.argument.impl.parsing.types.*;
+import com.diamondfire.helpbot.bot.command.argument.impl.types.GreedyStringArgument;
 import com.diamondfire.helpbot.bot.command.help.*;
 import com.diamondfire.helpbot.bot.command.impl.Command;
 import com.diamondfire.helpbot.bot.command.permissions.Permission;
@@ -38,7 +39,7 @@ public class EightBallCommand extends Command {
     public ArgumentSet compileArguments() {
         return new ArgumentSet()
                 .addArgument("question",
-                        new SingleArgumentContainer<>(new MessageArgument()));
+                        new SingleArgumentContainer<>(new GreedyStringArgument()));
     }
     
     @Override
