+++ src/main/java/com/diamondfire/helpbot/bot/command/impl/other/util/MimicCommand.java
@@ -1,7 +1,7 @@
 package com.diamondfire.helpbot.bot.command.impl.other.util;
 
 import com.diamondfire.helpbot.bot.command.argument.ArgumentSet;
-import com.diamondfire.helpbot.bot.command.argument.impl.parsing.types.MessageArgument;
+import com.diamondfire.helpbot.bot.command.argument.impl.types.GreedyStringArgument;
 import com.diamondfire.helpbot.bot.command.help.*;
 import com.diamondfire.helpbot.bot.command.impl.Command;
 import com.diamondfire.helpbot.bot.command.permissions.Permission;
@@ -28,7 +28,7 @@ public class MimicCommand extends Command {
     @Override
     public ArgumentSet compileArguments() {
         return new ArgumentSet()
-                .addArgument("msg", new MessageArgument());
+                .addArgument("msg", new GreedyStringArgument());
     }
     
     @Override
